css/css-fonts/parsing/font-size-adjust-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Level 3: parsing font-size-adjust with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-size-adjust-prop">
<meta name="assert" content="font-size-adjust supports the full grammar 'none | <number>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("font-size-adjust", "none");
test_valid_value("font-size-adjust", "10");
</script>
</body>
</html>
